feat: enforce RBAC across middleware, dashboard, and profiles
Add primaryOfficerVerifier and signedInVerifier to auth middleware.
Update ROUTES so officer/position management requires primary officer,
quote POST requires sign-in, and attendance endpoints require sign-in.
Register 18 previously unprotected API routes with correct verifiers.

Convert dashboard and mentoring layouts from client-side to server-side
auth enforcement. Add officer-only guards to users, alumni, and sponsors
pages so mentor-only users are scoped to the positions section. Gate
/profile/[id] behind owner-or-officer auth.

Fixes #332, fixes #333, fixes #334, fixes #335, fixes #336, fixes #337, fixes #338, fixes #339
