+++ next/app/(main)/dashboard/users/UsersPageContent.tsx
@@ -1,0 +1,11 @@
+"use client"
+
+import UsersSection from "../UsersSection"
+
+export default function UsersPageContent() {
+  return (
+    <div className="w-full max-w-7xl mx-auto p-4 sm:p-6 lg:p-8">
+      <UsersSection />
+    </div>
+  )
+}
