+++ next/app/(main)/dashboard/users/page.tsx
@@ -1,11 +1,13 @@
-"use client"
+import { redirect } from "next/navigation"
+import { getAuthLevel } from "@/lib/services/authLevelService"
+import UsersPageContent from "./UsersPageContent"
 
-import UsersSection from "../UsersSection"
+export default async function UsersPage() {
+  const authLevel = await getAuthLevel()
 
-export default function UsersPage() {
-  return (
-    <div className="w-full max-w-7xl mx-auto p-4 sm:p-6 lg:p-8">
-      <UsersSection />
-    </div>
-  )
+  if (!authLevel.isOfficer) {
+    redirect("/dashboard/positions")
+  }
+
+  return <UsersPageContent />
 }
