+++ next/lib/middlewares/authentication.ts
@@ -53,6 +53,21 @@ const mentorVerifier = authVerifierFactory((permissions) => {
 });
 
 /**
+ * Auth verifier that requires a primary officer (e.g. President, VP).
+ * Used for sensitive management operations like officer/position CRUD.
+ */
+const primaryOfficerVerifier = authVerifierFactory((permissions) => {
+  return { isAllowed: permissions.isPrimary, authType: "Primary Officer" };
+});
+
+/**
+ * Auth verifier that requires any signed-in user
+ */
+const signedInVerifier = authVerifierFactory((permissions) => {
+  return { isAllowed: permissions.isUser, authType: "Signed-in User" };
+});
+
+/**
  * Auth verifier specifically for the golinks route
  */
 const goLinkVerifier = async (request: NextRequest) => {
@@ -71,7 +86,7 @@ const goLinkVerifier = async (request: NextRequest) => {
 /**
  * Auth verifier for events:
  * - GET remains public
- * - attendance mutation endpoints keep their route-level auth logic
+ * - attendance mutation endpoints require signed-in user (route handles owner logic)
  * - all other non-GET event mutations require officer
  */
 const eventVerifier: AuthVerifier = async (request: NextRequest) => {
@@ -81,7 +96,7 @@ const eventVerifier: AuthVerifier = async (request: NextRequest) => {
   }
 
   if (/^\/api\/event\/[^/]+\/attendance$/.test(pathname)) {
-    return { isAllowed: true, authType: "None" };
+    return signedInVerifier(request);
   }
 
   return officerVerifier(request);
@@ -91,6 +106,11 @@ const eventVerifier: AuthVerifier = async (request: NextRequest) => {
  * An auth verifier that allows GET requests but makes sure all other requests are made by officers
  */
 const nonGetOfficerVerifier = nonGetVerifier(officerVerifier);
+
+/**
+ * An auth verifier that allows GET requests but requires primary officer for mutations
+ */
+const nonGetPrimaryOfficerVerifier = nonGetVerifier(primaryOfficerVerifier);
 
 /**
  * An auth verifier that allows GET requests but makes sure all other requests are made by mentors
@@ -123,35 +143,85 @@ const userVerifier: AuthVerifier = async (request: NextRequest) => {
 };
 
 /**
+ * Auth verifier for quotes:
+ * - GET is public (anyone can read quotes)
+ * - POST requires signed-in user (anyone logged in can submit)
+ * - PUT/DELETE pass through to route-level checks (self-edit with officer override)
+ */
+const quoteVerifier: AuthVerifier = async (request: NextRequest) => {
+  if (request.method === "GET") {
+    return { isAllowed: true, authType: "None" };
+  }
+  // POST, PUT, DELETE all require at least a signed-in user.
+  // PUT/DELETE route handlers enforce ownership or officer privilege.
+  return signedInVerifier(request);
+};
+
+/**
+ * Auth verifier for mentor applications:
+ * - GET passes through (route handles own-application vs manager logic)
+ * - POST requires signed-in user (anyone can apply)
+ * - PUT/PATCH/DELETE pass through (route enforces owner or mentoringHead/primary)
+ */
+const mentorApplicationVerifier: AuthVerifier = async (
+  request: NextRequest
+) => {
+  if (request.method === "GET") {
+    return { isAllowed: true, authType: "None" };
+  }
+  // All mutations require at least a signed-in user
+  return signedInVerifier(request);
+};
+
+/**
  * Map from API route name to authorization verifier. The verifier should be run against any request that
  * goes through that route.
  * Keys are the second element in the path segment; for example, the path "/api/golinks/officer" would
  * correspond to the key "golinks"
+ *
+ * IMPORTANT: Every API route directory must have an entry here. Routes without an entry
+ * will pass through without any auth check.
  */
 const ROUTES: { [key: string]: AuthVerifier } = {
   alumni: nonGetOfficerVerifier,
+  "alumni-candidates": officerVerifier,
   "alumni-requests": alumniRequestsVerifier,
+  aws: officerVerifier,
   calendar: nonGetOfficerVerifier,
   course: nonGetOfficerVerifier,
   courseTaken: nonGetMentorVerifier,
   departments: nonGetOfficerVerifier,
+  email: officerVerifier,
   event: eventVerifier,
   golinks: goLinkVerifier,
+  handover: nonGetPrimaryOfficerVerifier,
+  "headcount-import": primaryOfficerVerifier,
+  "headcount-trends": officerVerifier,
   hourBlocks: nonGetOfficerVerifier,
+  invitations: officerVerifier,
   memberships: nonGetOfficerVerifier,
+  "mentee-headcount": officerVerifier,
   mentor: nonGetOfficerVerifier,
+  "mentor-application": mentorApplicationVerifier,
+  "mentor-availability": nonGetMentorVerifier,
+  "mentor-semester": nonGetOfficerVerifier,
+  "mentoring-headcount": officerVerifier,
+  mentorSchedule: nonGetMentorVerifier,
   mentorSkill: nonGetMentorVerifier,
-  officer: nonGetOfficerVerifier,
-  "officer-positions": nonGetOfficerVerifier,
+  officer: nonGetPrimaryOfficerVerifier,
+  "officer-positions": nonGetPrimaryOfficerVerifier,
   project: nonGetOfficerVerifier,
   projectContributor: nonGetOfficerVerifier,
   purchasing: officerVerifier, // All purchasing routes require officer auth
-  quotes: nonGetOfficerVerifier,
+  quotes: quoteVerifier,
   schedule: nonGetMentorVerifier,
+  scheduleBlock: nonGetMentorVerifier,
   skills: nonGetOfficerVerifier,
   sponsor: nonGetOfficerVerifier,
+  "swipe-access": officerVerifier,
   user: userVerifier,
   userProject: nonGetOfficerVerifier,
+  when2meet: signedInVerifier,
 };
 
 const accessDenied = (authType: string, request: NextRequest) => {
